+++ kosh/ensemble.py
@@ -337,7 +337,7 @@ These datasets will inherit attributes and associated sources from the ensemble.
 
     @lock_strategies.lock_method
     def to_dataframe(self, data_columns=[], include_ensemble_attributes=True, include_ensemble_tags=True,
-                     *atts, **keys):
+                     typed_columns=None, *atts, **keys):
         """Return the find_datasets object as a Pandas DataFrame.
 
         Pass in the same arguments and keyword arguments as the find method.
@@ -362,6 +362,39 @@ These datasets will inherit attributes and associated sources from the ensemble.
         :type include_ensemble_attributes: bool, optional
         :param include_ensemble_tags: Include ensemble tags in DataFrame.
         :type include_ensemble_tags: bool, optional
+        :param typed_columns: Dictionary of column types and replacement values.
+            If omitted, MySQL defaults are returned.
+
+            Supported types:
+                [
+                    "Int8", "Int16", "Int32", "Int64",
+                    "UInt8", "UInt16", "UInt32", "UInt64",
+                    "float16", "float32", "float64",
+                    "Float32", "Float64",
+                    "bool", "boolean", "string", "object",
+                    "category", "datetime64[ns]", "timedelta64[ns]",
+                ]
+
+            Example:
+                {
+                    "my_int": {
+                        "type": "Int64",
+                        "replace": {None: 42},
+                    },
+                    "my_float": {
+                        "type": "float64",
+                        "replace": {None: np.nan, 999: -1},
+                    },
+                    "my_boolean": {
+                        "type": "boolean",
+                        "replace": {None: False},
+                    },
+                    "my_string": {
+                        "type": "string",
+                        "replace": {"This is empty": "Nothing Here"},
+                    },
+                }
+        :type typed_columns: dict, optional
         :return: Pandas DataFrame
         :rtype: Pandas DataFrame
         """
@@ -425,4 +458,14 @@ These datasets will inherit attributes and associated sources from the ensemble.
                 attr_dict[f"{self.id}_ENSEMBLE_ATTRIBUTE_{key}"] = [val] * total_datasets
 
         df = pd.DataFrame(attr_dict)
+
+        if isinstance(typed_columns, dict):
+            for key, config in typed_columns.items():
+                series = df[key]
+
+                if "replace" in config:
+                    series = series.replace(config["replace"]).infer_objects(copy=False)
+
+                df[key] = series.astype(config["type"])
+
         return df